compiles again

+++ src/Test.java
@@ -1,6 +1,7 @@
 import com.peter.crypto.numberfield2d.NumberField;
 import com.peter.crypto.numberfield2d.NumberFieldFactory;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 
@@ -9,7 +10,7 @@ public class Test
     public static void main (String[] args) throws Exception
     {
         NumberField n =
-                NumberFieldFactory.countedSpiralRightDown (Double.class, 5, "motherfucker");
+                NumberFieldFactory.countedSpiralRightDown (BigDecimal.class, 5, "motherfucker");
         System.out.println(n);
 
 //        BigInteger bi = PiDigitsHex.packedPi(0,1000);
